refactor(auth): move sign up related widgets to sign_up folder and move login related widgets to login folder
