+++ lib/features/auth/presentation/views/widgets/login/login_view_body.dart
@@ -2,9 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:nova_store_app/core/utils/spacing.dart';
 import 'package:nova_store_app/features/auth/presentation/views/widgets/auth_app_bar.dart';
-import 'package:nova_store_app/features/auth/presentation/views/widgets/dont_have_an_account.dart';
+import 'package:nova_store_app/features/auth/presentation/views/widgets/login/dont_have_an_account.dart';
 import 'package:nova_store_app/features/auth/presentation/views/widgets/enter_email_section.dart';
-import 'package:nova_store_app/features/auth/presentation/views/widgets/login_button.dart';
+import 'package:nova_store_app/features/auth/presentation/views/widgets/login/login_button.dart';
 
 class LoginViewBody extends StatefulWidget {
   const LoginViewBody({super.key});
