+++ lib/features/auth/presentation/views/widgets/sign_up/sign_up_button_builder.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:nova_store_app/core/widgets/custom_loading_button.dart';
 import 'package:nova_store_app/features/auth/presentation/manager/sign_up_cubit/sign_up_cubit.dart';
-import 'package:nova_store_app/features/auth/presentation/views/widgets/sign_up_button.dart';
+import 'package:nova_store_app/features/auth/presentation/views/widgets/sign_up/sign_up_button.dart';
 
 class SignUpButtonBuilder extends StatelessWidget {
   const SignUpButtonBuilder({
