+++ lib/features/auth/presentation/views/widgets/sign_up/sign_up_view_body.dart
@@ -3,10 +3,10 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:nova_store_app/core/utils/spacing.dart';
 import 'package:nova_store_app/features/auth/presentation/manager/sign_up_cubit/sign_up_cubit.dart';
-import 'package:nova_store_app/features/auth/presentation/views/widgets/already_have_an_account.dart';
+import 'package:nova_store_app/features/auth/presentation/views/widgets/sign_up/already_have_an_account.dart';
 import 'package:nova_store_app/features/auth/presentation/views/widgets/auth_app_bar.dart';
 import 'package:nova_store_app/features/auth/presentation/views/widgets/enter_email_section.dart';
-import 'package:nova_store_app/features/auth/presentation/views/widgets/sign_up_button_builder.dart';
+import 'package:nova_store_app/features/auth/presentation/views/widgets/sign_up/sign_up_button_builder.dart';
 
 class SignUpViewBody extends StatefulWidget {
   const SignUpViewBody({super.key});
